requirements.txt
# Core LLM and AI libraries
google-genai>=1.0.0
openai>=1.0.0  # For OpenRouter compatibility

# LangChain (minimal usage - only for text splitting and document abstraction)
langchain>=0.1.0
langchain-community>=0.0.20

# Vector store
faiss-cpu>=1.7.4

# PDF processing
pdfplumber>=0.10.0
pypdf>=3.17.0

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6  # For file uploads

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP client for OpenRouter
httpx>=0.25.0
